+++ grid2op/Reward/_AlarmScore.py
@@ -7,11 +7,12 @@
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
 import numpy as np
+import copy
 
 from grid2op.Exceptions import Grid2OpException
 from grid2op.Reward import AlarmReward
 from grid2op.dtypes import dt_float
-import copy
+
 
 class _AlarmScore(AlarmReward):
     """
@@ -56,19 +57,25 @@ class _AlarmScore(AlarmReward):
         # required if you want to design a custom reward taking into account the
         # alarm feature
         self.reward_min = dt_float(-2.)
-        #we keep other parameters values from AlarmReward as is
+        # we keep other parameters values from AlarmReward as is
 
         self.mult_for_right_zone = 1.5
         self.window_disconnection = 4
 
         self.disc_lines_all_before_cascade = []
+        self.n_line = None
+
+        # This class remembers the past state of the grid, this does not make sense for the "simulate" env
+        # so i deactivate it in this case.
+        from grid2op.Observation._ObsEnv import _ObsEnv  # to avoid circular dependencies
+        self._deactivate_reward_cls = (_ObsEnv, )
 
     def initialize(self, env):
         if not env._has_attention_budget:
-            raise Grid2OpException("Impossible to use the \"AlarmReward\" with an environment for which this feature "
+            raise Grid2OpException("Impossible to use the \"_AlarmScore\" with an environment for which this feature "
                                    "is disabled. Please make sure \"env._has_attention_budget\" is set to ``True`` or "
                                    "change the reward class with `grid2op.make(..., reward_class=AnyOtherReward)`")
-        self.n_line=env.n_line
+        self.n_line = env.n_line
         self.reset(env)
 
     def reset(self, env):
@@ -78,31 +85,33 @@ class _AlarmScore(AlarmReward):
 
     def _lines_disconnected_first(self, disc_lines_at_cascading_time):
         """
-        here we detect the disconnected lines that we will consider to compute the mult_for_zone multiplifying factor.
+        here we detect the disconnected lines that we will consider to compute the `mult_for_zone` multiplying factor.
         Either the lines that were disconnected in a short period before final failure. Otherwise the first lines
-        disconnedted at the time of failure
+        disconnected at the time of failure
 
         :param disc_lines_at_cascading_time: lines that are disconnected first at time of failure
         :return:
         """
 
-        disc_lines_to_consider_for_score = np.zeros(self.n_line,dtype=bool)
+        disc_lines_to_consider_for_score = np.zeros(self.n_line, dtype=bool)
 
         nb_obs = len(self.disc_lines_all_before_cascade)
 
-        for t in range(nb_obs - self.window_disconnection, nb_obs):
-            disc_lines_to_consider_for_score[self.disc_lines_all_before_cascade[t] >= 0]=True
+        for step in range(nb_obs - self.window_disconnection, nb_obs):
+            disc_lines_to_consider_for_score[self.disc_lines_all_before_cascade[step] >= 0] = True
 
-        if (np.sum(disc_lines_to_consider_for_score) == 0):
-            disc_lines_to_consider_for_score = disc_lines_at_cascading_time==0
+        if np.sum(disc_lines_to_consider_for_score) == 0:
+            disc_lines_to_consider_for_score = disc_lines_at_cascading_time == 0
 
-        #if we are there, it is because we have identified before that the failure is due to disconnected powerlines
-        assert(np.any(disc_lines_to_consider_for_score))
+        # if we are there, it is because we have identified before that the failure is due to disconnected powerlines
+        assert np.any(disc_lines_to_consider_for_score)
 
-        #we transform the vector so that diconnected lines have a zero, to be coherent with env._disc_lines
-        return 1-disc_lines_to_consider_for_score
+        # we transform the vector so that disconnected lines have a zero, to be coherent with env._disc_lines
+        return 1 - disc_lines_to_consider_for_score
 
     def __call__(self, action, env, has_error, is_done, is_illegal, is_ambiguous):
+        if isinstance(env, self._deactivate_reward_cls):
+            return self.reward_no_game_over
 
         disc_lines_now = env._disc_lines
 
@@ -110,7 +119,6 @@ class _AlarmScore(AlarmReward):
             if not has_error:
                 # agent went until the end
                 return self.reward_max
-
             if np.all(env._disc_lines == -1):
                 # game over is not caused by the tripping of a powerline
                 return self.reward_min
@@ -139,6 +147,8 @@ class _AlarmScore(AlarmReward):
             return best_score
         else:
             # make sure to deepcopy, otherwise it gets updated with the last timestep value for every previous timesteps
-            self.disc_lines_all_before_cascade.append(copy.deepcopy(disc_lines_now))  # we log the line disconnected over time
+            # we log the line disconnected over time
+            # TODO have a cache there and store only the last few states, most of what is stored here is not used
+            self.disc_lines_all_before_cascade.append(copy.deepcopy(disc_lines_now))
             res = self.reward_no_game_over
         return res