+++ CHANGELOG.rst
@@ -23,6 +23,11 @@ Change Log
 - [???] "asynch" multienv
 - [???] properly model interconnecting powerlines
 
+[1.6.2] (hotfix) - 2021-08-18
+-----------------------------
+- [FIXED] an issue when using `obs.simulate` with `_AlarmScore` (major bug)
+- [FIXED] now properly initialized the "complete_action_class" of the backend (minor bug)
+
 [1.6.2] - 2021-07-27
 ---------------------
 - [ADDED] the complete support for pickling grid2op classes. This is a major feature that allows to use grid2op